Make SiegeWar use newer StatusScreen objects.
- SiegeWar now uses the StatusScreen objects to add details to Towny
status screens.
- Includes:
  - Nation Refund, Plunder, Military Salary details are now
click-to-collect from the Resident Status Screen.
  - Town status screen moves many Siege details to [Sieged] hoverable
component.
  - Nation status screen only shows lists when they are greater than
zero.
  - Lists of Towns in the Nation status screens will now condense to
shorter, un-formatted names when the lists are longer than 20.
    - Partially completes #413.
