+++ src/main/resources/english.yml
@@ -492,3 +492,8 @@ attacker_has_begun_a_siegecamp_session: '%s has begun a SiegeCamp session agains
 msg_bossbar_preference_set: "&bBossBar preference set to %s."
 bossbar_msg_battle_time_remaining: '&2Battle Session Time Remaining: &a%s'
 msg_err_siege_war_revolt_immunity_permanent: "&cYour town has permanent revolt immunity. It cannot revolt."
+status_sieged: "%s[%sSieged%s]"
+status_nation_refund: "Nation Refund: "
+status_plunder: "Plunder: "
+status_military_salary: "Military Salary: "
+hover_message_click_to_claim: "Click to claim money owed to you."
